Updated wishController and dateutils for accurate date calculation and edge cases

+++ backend/utils/dateUtils.js
@@ -66,7 +66,7 @@ export const isToday = (date) => {
 export const formatDate = (date) => {
   return new Date(date).toLocaleDateString('en-US', {
     year: 'numeric',
-    month: 'short',
+    month: 'long',
     day: 'numeric'
   });
 };
